# backend-api/getdataphp/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Đăng Nhập",
        "description": "Tài liệu API cho ứng dụng Mobile",
        "version": "1.0.0"
    },
    "paths": {
        "/api/dsdonvi": {
            "get": {
                "tags": [
                    "Administrative"
                ],
                "summary": "Lấy danh sách đơn vị",
                "description": "Trả về toàn bộ danh sách các đơn vị",
                "operationId": "5d616d22c2185e7012fb5e07415756c8",
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/dsdiaban": {
            "get": {
                "tags": [
                    "Administrative"
                ],
                "summary": "Lấy danh sách địa bàn",
                "description": "Trả về toàn bộ danh sách các địa bàn hành chính",
                "operationId": "d2aa1e9616e7943103ea7448ec38b905",
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Đăng nhập hệ thống",
                "description": "Cho phép người dùng đã có tài khoản trong SQL Server đăng nhập",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đăng nhập thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đăng nhập thành công"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Sai tên đăng nhập hoặc mật khẩu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tên đăng nhập hoặc mật khẩu không đúng"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/getUsers": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Lấy thông tin người dùng",
                "description": "Truy xuất thông tin chi tiết của người dùng bằng username",
                "operationId": "200f43e2a214715fabc23acd8ec85931",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "Tên đăng nhập của người dùng",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy người dùng"
                    }
                }
            }
        },
        "/api/thuetainguyen": {
            "get": {
                "tags": [
                    "ThueTaiNguyen"
                ],
                "summary": "Lấy danh sách Thuế Tài Nguyên",
                "description": "Trả về danh sách phân trang các hồ sơ thuế tài nguyên",
                "operationId": "12f910508f36770c78f199cb5284e84b",
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/thuetainguyen/{mahs}": {
            "get": {
                "tags": [
                    "ThueTaiNguyen"
                ],
                "summary": "Lấy chi tiết Thuế Tài Nguyên",
                "description": "Trả về thông tin chi tiết và danh sách chi tiết (thuetainguyenct) theo mã hồ sơ",
                "operationId": "c8d52b931f93ff1f0c8ea90a80322030",
                "parameters": [
                    {
                        "name": "mahs",
                        "in": "path",
                        "description": "Mã hồ sơ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy hồ sơ"
                    }
                }
            }
        },
        "/api/thuetainguyen/stats": {
            "get": {
                "tags": [
                    "ThueTaiNguyen"
                ],
                "summary": "Lấy thống kê hồ sơ Thuế Tài Nguyên",
                "description": "Trả về số lượng tổng cộng, hồ sơ đã hoàn thành (HT) và chưa hoàn thành. Hỗ trợ lọc theo thời gian.",
                "operationId": "7ac72ade8c6cbe742a1fbf8f54326583",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Lọc theo thời gian: all, week, month, quarter, year",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "all"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "completed": {
                                                    "type": "integer"
                                                },
                                                "incomplete": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}